Modification to Component Panel and formerly teamNameInput. TeamNameInput now changed to TeamsConfirmed, added all team customisations to one display panel
